refactor(channels): split compact command wrappers
